refactoring : #7 Items(일급 컬렉션), ItemInfo(VO 명칭 변경)
