+++ src/main/java/me/hero/myrefrigerator/common/vo/Items.java
@@ -1,0 +1,13 @@
+package me.hero.myrefrigerator.common.vo;
+
+import me.hero.myrefrigerator.domain.Item;
+
+import java.util.List;
+
+public class Items {
+    private List<Item> items;
+
+    public Items(List<Item> items) {
+        this.items = items;
+    }
+}
