+++ src/main/java/me/hero/myrefrigerator/domain/dto/ItemInfo.java
@@ -2,10 +2,9 @@ package me.hero.myrefrigerator.domain.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @AllArgsConstructor @Data
-public class ItemResponse {
+public class ItemInfo {
     private Long id;
     private String refrigeratorName;
     private String name;
