+++ src/main/java/me/hero/myrefrigerator/repository/RefrigeratorQueryRepositoryImpl.java
@@ -2,14 +2,11 @@ package me.hero.myrefrigerator.repository;
 
 import lombok.RequiredArgsConstructor;
 import me.hero.myrefrigerator.domain.Item;
-import me.hero.myrefrigerator.domain.Refrigerator;
-import me.hero.myrefrigerator.domain.dto.ItemResponse;
-import me.hero.myrefrigerator.repository.mybatis.RefrigeratorMapper;
+import me.hero.myrefrigerator.domain.dto.ItemInfo;
 import me.hero.myrefrigerator.repository.mybatis.RefrigeratorQueryMapper;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 @RequiredArgsConstructor
@@ -23,7 +20,7 @@ public class RefrigeratorQueryRepositoryImpl implements RefrigeratorQueryReposit
 
 
     @Override
-    public List<ItemResponse> findItemsOrderByRefrigerator() {
+    public List<ItemInfo> findItemsOrderByRefrigerator() {
         return refrigeratorQueryMapper.findItemsOrderByRefrigerator();
     }
 }
